refactor: extract peer log helpers into internal/log and share one Auth
- Move PeerAddr/LogContext from internal/service/auth to the shared
  internal/log package so peer-enriched logging no longer requires
  importing the auth package; update the gRPC interceptors, the router
  stream handler, and the tests accordingly.
- Store a single memoized *auth.Auth on ControllerService (sync.Once)
  instead of constructing one per RPC, matching ClientService.
- Rework TestAuthClient_NoTokenLeak/TestAuthExporter_NoTokenLeak to
  inject the sensitive token as real incoming metadata through the
  production BearerTokenAuthenticator and assert the verifier actually
  received it, so the no-leak assertions can fail on a regression;
  previously the stub returned a fixed generic error, so the token
  never entered the code path and the assertions were tautological.
  Drop TestAuthClient_ErrorLogIncludesNoDuplicateTokenInMessage, whose
  only assertion (no "Bearer" in logs) is now covered by both rewritten
  tests with a real bearer header present.
