+++ controller/internal/service/auth/auth.go
@@ -2,43 +2,17 @@ package auth
 
 import (
 	"context"
-	"net"
 
 	jumpstarterdevv1alpha1 "github.com/jumpstarter-dev/jumpstarter-controller/api/v1alpha1"
 	"github.com/jumpstarter-dev/jumpstarter-controller/internal/authentication"
 	"github.com/jumpstarter-dev/jumpstarter-controller/internal/authorization"
 	"github.com/jumpstarter-dev/jumpstarter-controller/internal/oidc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/status"
 	"k8s.io/apiserver/pkg/authorization/authorizer"
 	kclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
-
-// PeerAddr returns the remote IP address from the gRPC peer info stored in ctx,
-// or "unknown" if unavailable. Port number and transport paths (e.g. Unix socket
-// paths) are intentionally stripped to avoid leaking internal details.
-func PeerAddr(ctx context.Context) string {
-	p, ok := peer.FromContext(ctx)
-	if !ok || p.Addr == nil {
-		return "unknown"
-	}
-	host, _, err := net.SplitHostPort(p.Addr.String())
-	if err != nil {
-		return "unknown"
-	}
-	return host
-}
-
-// LogContext returns ctx with its logger enriched with the peer address under
-// the "peer" key ("unknown" when no usable address is available, so all log
-// lines share a uniform shape). The gRPC interceptors apply this to every RPC;
-// it owns the peer enrichment, so loggers derived from ctx (including the
-// auth-failure logs below) must not add "peer" again.
-func LogContext(ctx context.Context) context.Context {
-	return log.IntoContext(ctx, log.FromContext(ctx, "peer", PeerAddr(ctx)))
-}
 
 type Auth struct {
 	client kclient.Client
@@ -64,7 +38,8 @@ func NewAuth(
 // VerifyClient authenticates the client token in ctx and returns the matching
 // Client object without enforcing a namespace. Authentication failures are
 // logged via the context logger, which carries the peer address when the
-// caller applied LogContext (as the gRPC interceptors do).
+// caller applied the shared log package's LogContext (as the gRPC
+// interceptors do).
 func (s *Auth) VerifyClient(ctx context.Context) (*jumpstarterdevv1alpha1.Client, error) {
 	jclient, err := oidc.VerifyClientObjectToken(
 		ctx,
@@ -100,7 +75,8 @@ func (s *Auth) AuthClient(ctx context.Context, namespace string) (*jumpstarterde
 // VerifyExporter authenticates the exporter token in ctx and returns the
 // matching Exporter object without enforcing a namespace. Authentication
 // failures are logged via the context logger, which carries the peer address
-// when the caller applied LogContext (as the gRPC interceptors do).
+// when the caller applied the shared log package's LogContext (as the gRPC
+// interceptors do).
 func (s *Auth) VerifyExporter(ctx context.Context) (*jumpstarterdevv1alpha1.Exporter, error) {
 	jexporter, err := oidc.VerifyExporterObjectToken(
 		ctx,
