+++ controller/internal/service/controller_service.go
@@ -85,6 +85,8 @@ type ControllerService struct {
 	Signer        *oidc.Signer
 	listenQueues  sync.Map
 	leaseLocks    sync.Map
+	authOnce      sync.Once
+	auth          *auth.Auth
 }
 
 type listenQueue struct {
@@ -206,26 +208,30 @@ type wrappedStream struct {
 	grpc.ServerStream
 }
 
-// logContext enriches the context logger with the peer address so every log
-// line (including auth failures logged by the auth package) carries a "peer"
-// key. See auth.LogContext for the enrichment semantics.
-func logContext(ctx context.Context) context.Context {
-	return auth.LogContext(ctx)
+func (w *wrappedStream) Context() context.Context {
+	return jlog.LogContext(w.ServerStream.Context())
 }
 
-func (w *wrappedStream) Context() context.Context {
-	return logContext(w.ServerStream.Context())
+// getAuth lazily builds the shared *auth.Auth from the service's immutable
+// dependencies on first use. Lazy init (rather than init in Start) keeps bare
+// &ControllerService{...} construction working, which tests rely on to call
+// RPC methods without Start; sync.Once makes the first concurrent RPCs safe.
+func (s *ControllerService) getAuth() *auth.Auth {
+	s.authOnce.Do(func() {
+		s.auth = auth.NewAuth(s.Client, s.Authn, s.Authz, s.Attr)
+	})
+	return s.auth
 }
 
 // authenticateClient and authenticateExporter delegate to the auth package,
 // which owns auth-failure logging (with peer address) for all services.
 
 func (s *ControllerService) authenticateClient(ctx context.Context) (*jumpstarterdevv1alpha1.Client, error) {
-	return auth.NewAuth(s.Client, s.Authn, s.Authz, s.Attr).VerifyClient(ctx)
+	return s.getAuth().VerifyClient(ctx)
 }
 
 func (s *ControllerService) authenticateExporter(ctx context.Context) (*jumpstarterdevv1alpha1.Exporter, error) {
-	return auth.NewAuth(s.Client, s.Authn, s.Authz, s.Attr).VerifyExporter(ctx)
+	return s.getAuth().VerifyExporter(ctx)
 }
 
 func (s *ControllerService) Register(ctx context.Context, req *pb.RegisterRequest) (*pb.RegisterResponse, error) {
@@ -1116,7 +1122,7 @@ func (s *ControllerService) Start(ctx context.Context) error {
 			_ *grpc.UnaryServerInfo,
 			handler grpc.UnaryHandler,
 		) (resp any, err error) {
-			return handler(logContext(gctx), req)
+			return handler(jlog.LogContext(gctx), req)
 		}, recovery.UnaryServerInterceptor()),
 		grpc.ChainStreamInterceptor(func(
 			srv any,
@@ -1132,7 +1138,7 @@ func (s *ControllerService) Start(ctx context.Context) error {
 	pb.RegisterControllerServiceServer(server, s)
 	cpb.RegisterClientServiceServer(
 		server,
-		clientsvcv1.NewClientService(s.Client, *auth.NewAuth(s.Client, s.Authn, s.Authz, s.Attr), s.effectiveMaxTags(), s.Signer),
+		clientsvcv1.NewClientService(s.Client, *s.getAuth(), s.effectiveMaxTags(), s.Signer),
 	)
 
 	// Register the standard gRPC health checking service (grpc.health.v1.Health).
